ppe/requirements.txt
# ==== Core Computer Vision Stack ====
ultralytics>=8.0.20
deep_sort_realtime==1.3.2
opencv-python==4.7.0.72
numpy==1.24.4
scikit-image==0.19.3
Pillow==9.5.0

torch==2.1.2+cu121
torchvision==0.16.2+cu121
torchaudio==2.1.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# ==== xformers ====
xformers==0.0.23.post1

# ==== FastAPI & WebRTC Stack ====
fastapi==0.116.1
uvicorn==0.35.0
aiortc==1.13.0
av==14.4.0
python-dotenv==1.0.1
psutil==7.0.0

# ==== Utilities ====
pydantic==2.11.1
starlette==0.47.2
pydantic-settings==2.7.0   

# ==== Database ====
SQLAlchemy==2.0.34
PyMySQL==1.1.1

# ==== BoT-SORT Dependency ====
lap>=0.5.12
